Extract condition evaluation to class for testing
