+++ test/unit/index.spec.ts
@@ -97,6 +97,7 @@ import './security-rules/security-rules-api-client.spec';
 import './remote-config/index.spec';
 import './remote-config/remote-config.spec';
 import './remote-config/remote-config-api-client.spec';
+import './remote-config/remote-config-condition-evaluator.spec';
 
 // AppCheck
 import './app-check/app-check.spec';
