+++ test/unit/remote-config/remote-config.spec.ts
@@ -201,83 +201,6 @@ describe('RemoteConfig', () => {
     stubs = [];
   });
 
-  describe('evaluateCondition', () => {
-    it('should evaluate empty OR condition to false', () => {
-      const condition = {
-        name: 'is_enabled',
-        or: {
-        }
-      };
-      const template = remoteConfig.initServerTemplate()
-      expect(template.evaluateCondition(condition, 0)).to.be.false;
-    });
-
-    it('should evaluate empty OR.AND condition to true', () => {
-      const condition = {
-        name: 'is_enabled',
-        or: {
-          conditions: [
-            {
-              name: '', // Note we should differentiate named from unnamed conditions
-              and: {
-              }
-            }
-          ]
-        }
-      };
-      const template = remoteConfig.initServerTemplate()
-      expect(template.evaluateCondition(condition, 0)).to.be.true;
-    });
-
-    it('should evaluate OR.AND.TRUE condition to true', () => {
-      const condition = {
-        name: 'is_enabled',
-        or: {
-          conditions: [
-            {
-              name: '', // Note we should differentiate named from unnamed conditions
-              and: {
-                conditions: [
-                  {
-                    name: '',
-                    true: {
-                    }
-                  }
-                ]
-              }
-            }
-          ]
-        }
-      };
-      const template = remoteConfig.initServerTemplate()
-      expect(template.evaluateCondition(condition, 0)).to.be.true;
-    });
-
-    it('should evaluate OR.AND.FALSE condition to false', () => {
-      const condition = {
-        name: 'is_enabled',
-        or: {
-          conditions: [
-            {
-              name: '', // Note we should differentiate named from unnamed conditions
-              and: {
-                conditions: [
-                  {
-                    name: '',
-                    false: {
-                    }
-                  }
-                ]
-              }
-            }
-          ]
-        }
-      };
-      const template = remoteConfig.initServerTemplate()
-      expect(template.evaluateCondition(condition, 0)).to.be.false;
-    });
-  });
-
   describe('Constructor', () => {
     const invalidApps = [null, NaN, 0, 1, true, false, '', 'a', [], [1, 'a'], {}, { a: 1 }, _.noop];
     invalidApps.forEach((invalidApp) => {
